Update and fix OpenGL sample project 'Minimal GLFW project with VBO'
The hyperlink used by the CmakeFile.txt for automating the downloading
of glew library, that is used for loading OpenGL function pointers on
Windows, no longer works. As a result, the hyperlink was updated and a
new procedure was added. All three versions of the CMakeLists.txt file
were fixed and updated with the new glew hyperlink.

No longer works:
  + https://megalink.dl.sourceforge.net/project/glew/glew/2.1.0/glew-2.1.0-win32.zip

Link replaced with:
  + https://ufpr.dl.sourceforge.net/project/glew/glew/2.1.0/glew-2.1.0-win32.zip